LayoutTests/imported/w3c/web-platform-tests/cookies/prefix/__secure.header.https-expected.txt


PASS __Secure: secure origin: Should not set 'Path=/;'
FAIL __SeCuRe: secure origin: Should not set 'Path=/;' assert_equals: expected (undefined) undefined but got (string) "0.9828188170834881"
PASS __Secure: secure origin: Should set 'Secure;Path=/;'
PASS __SeCuRe: secure origin: Should set 'Secure;Path=/;'
PASS __Secure: secure origin: Should not set 'Path=/;MaxAge=10'
FAIL __SeCuRe: secure origin: Should not set 'Path=/;MaxAge=10' assert_equals: expected (undefined) undefined but got (string) "0.42379718720806825"
PASS __Secure: secure origin: Should set 'Secure;Path=/;MaxAge=10'
PASS __SeCuRe: secure origin: Should set 'Secure;Path=/;MaxAge=10'
PASS __Secure: secure origin: Should not set 'Path=/;HttpOnly'
FAIL __SeCuRe: secure origin: Should not set 'Path=/;HttpOnly' assert_equals: expected (undefined) undefined but got (string) "0.811318163826578"
PASS __Secure: secure origin: Should set 'Secure;Path=/;HttpOnly'
PASS __SeCuRe: secure origin: Should set 'Secure;Path=/;HttpOnly'
PASS __Secure: secure origin: Should not set 'Path=/;domain=not-web-platform.test'
PASS __SeCuRe: secure origin: Should not set 'Path=/;domain=not-web-platform.test'
FAIL __Secure: secure origin: Should set 'Secure;Path=/;domain=not-web-platform.test' assert_equals: expected (string) "0.21745972996883733" but got (undefined) undefined
FAIL __SeCuRe: secure origin: Should set 'Secure;Path=/;domain=not-web-platform.test' assert_equals: expected (string) "0.7893561725008221" but got (undefined) undefined
